continue refactor, create common_fw, and move hardware specific declarations away from it. Need to think more about how to implement struct array indexing. It's too simple to have to go in common fw w/ so much data being passed to it.
