+++ platformio.ini
@@ -8,12 +8,11 @@
 ; Please visit documentation for the other options and examples
 ; https://docs.platformio.org/page/projectconf.html
 
-[env:E1]
+[env]
 platform = teensy
 board = teensy41
 framework = arduino
-src_filter = +<E1_coldflow.cpp>
-lib_deps = 
+lib_deps =
 	adafruit/Adafruit BusIO@^1.7.1
 	adafruit/Adafruit MCP9600 Library@^1.1.1
 	adafruit/Adafruit GPS Library@^1.5.3
@@ -22,8 +21,8 @@ lib_deps =
 	greiman/SdFat@^2.0.4
 	adafruit/Adafruit BMP3XX Library@^2.0.1
 
-[env:E3]
-platform = teensy
-board = teensy41
-framework = arduino
-src_filter = +<E3.cpp>
+[env:E1]
+src_filter = +<E1_coldflow/> -<LAD4_1/>
+
+[env:lad4_1]
+src_filter = +<LAD4_1/> -<E1_coldflow/>
